move test_engine.py to tests folder
