+++ .gitignore
@@ -29,4 +29,4 @@ lobe_cp
 mri_cp
 spleen_cp
 lobe_cp_sliding_window
-*.json
+.vscode/
