+++ liver_imaging_analysis/engine/engine.py
@@ -150,7 +150,7 @@ class Engine:
         Raises:
             NotImplementedError: When the function is not implemented.
         """
-        return Compose([])
+        raise NotImplementedError
 
     def get_pretesting_transforms(self, *args, **kwargs):
         """
@@ -161,7 +161,7 @@ class Engine:
         Raises:
             NotImplementedError: When the function is not implemented.
         """
-        return Compose([])
+        raise NotImplementedError
 
     def get_postprocessing_transforms(self, *args, **kwargs):
         """
